update R308 Lib

+++ lib/CTB_R308/CTB_R308.cpp
@@ -44,7 +44,7 @@ bool R308::init() {
             2:No finger.传感器无手指
             3:Scan Failed.录入失败
 */
-short R308::cmdGetImg() {
+int8_t R308::cmdGetImg() {
   serialClean();
   SerPort.write(&packHead[0], 6);
   SerPort.write(&packGetImg[0], 6);
@@ -65,7 +65,7 @@ short R308::cmdGetImg() {
             7:Few features.特征点过少
             15:No image in buffer.没有图像
 */
-short R308::cmdToBuffer1() {
+int8_t R308::cmdToBuffer1() {
   serialClean();
   SerPort.write(&packHead[0], 6);
   SerPort.write(&packToBuffer1[0], 7);
@@ -86,7 +86,7 @@ short R308::cmdToBuffer1() {
             7:Few features.特征点过少
             15:No image in buffer.没有图像
 */
-short R308::cmdToBuffer2() {
+int8_t R308::cmdToBuffer2() {
   serialClean();
   SerPort.write(&packHead[0], 6);
   SerPort.write(&packToBuffer2[0], 7);
@@ -105,7 +105,7 @@ short R308::cmdToBuffer2() {
             1:Pack error.收包有误
             A:Merge error:Not same finger.合并错误:非同一手指
 */
-short R308::cmdRegModel() {
+int8_t R308::cmdRegModel() {
   serialClean();
   SerPort.write(&packHead[0], 6);
   SerPort.write(&packRegModel[0], 6);
@@ -124,7 +124,7 @@ short R308::cmdRegModel() {
             1:Pack error.收包有误
             11:Empty failed.清空失败
 */
-short R308::cmdEmpty() {
+int8_t R308::cmdEmpty() {
   serialClean();
   SerPort.write(&packHead[0], 6);
   SerPort.write(&packEmpty[0], 6);
@@ -145,14 +145,14 @@ short R308::cmdEmpty() {
             B:PageID out of range.超出指纹库范围
             18:Flash error.写Flash出错
 */
-short R308::cmdSaveFinger(uint8_t bufferID, uint16_t pageID) {
+int8_t R308::cmdSaveFinger(uint8_t bufferID, uint16_t pageID) {
   volatile unsigned int Sum = 0;
 
   packSaveFinger[4] = bufferID;
   packSaveFinger[5] = (pageID & 0xFF00) >> 8;
   packSaveFinger[6] = pageID & 0x00FF;
 
-  for (short i = 0; i < 7; i++)
+  for (int8_t i = 0; i < 7; i++)
     Sum += packSaveFinger[i];
 
   packSaveFinger[7] = (Sum & 0xFF00) >> 8;
@@ -177,8 +177,8 @@ short R308::cmdSaveFinger(uint8_t bufferID, uint16_t pageID) {
             1:Pack error.收包有误
             9:Nothing matched.未搜索到
 */
-short R308::cmdSearch(uint8_t bufferID, uint16_t startPageID,
-                      uint16_t pageNum) {
+int8_t R308::cmdSearch(uint8_t bufferID, uint16_t startPageID,
+                       uint16_t pageNum) {
   volatile unsigned int Sum = 0;
 
   packSearch[4] = bufferID;
@@ -187,7 +187,7 @@ short R308::cmdSearch(uint8_t bufferID, uint16_t startPageID,
   packSearch[7] = (pageNum & 0xFF00) >> 8;
   packSearch[8] = pageNum & 0x00FF;
 
-  for (short i = 0; i < 9; i++)
+  for (int8_t i = 0; i < 9; i++)
     Sum += packSearch[i];
 
   packSearch[9] = (Sum & 0xFF00) >> 8;
@@ -212,7 +212,7 @@ short R308::cmdSearch(uint8_t bufferID, uint16_t startPageID,
             1:Pack error.收包有误
             10:Delete failed.删除失败
 */
-short R308::cmdDeleteModel(uint16_t startPageID, uint16_t pageNum) {
+int8_t R308::cmdDeleteModel(uint16_t startPageID, uint16_t pageNum) {
   volatile unsigned int Sum = 0;
 
   packDeleteModel[4] = (startPageID & 0xFF00) >> 8;
@@ -220,7 +220,7 @@ short R308::cmdDeleteModel(uint16_t startPageID, uint16_t pageNum) {
   packDeleteModel[6] = (pageNum & 0xFF00) >> 8;
   packDeleteModel[7] = pageNum & 0x00FF;
 
-  for (short i = 0; i < 8; i++)
+  for (int8_t i = 0; i < 8; i++)
     Sum += packDeleteModel[i];
 
   packDeleteModel[8] = (Sum & 0xFF00) >> 8;
@@ -241,7 +241,7 @@ short R308::cmdDeleteModel(uint16_t startPageID, uint16_t pageNum) {
    \brief 清空串口缓存
 */
 void R308::serialClean() {
-  for (short i = 0; i < 10; i++) {
+  for (int8_t i = 0; i < 10; i++) {
     packSerialRead[i] = 0xFF;
   }
   while (SerPort.read() >= 0) {
@@ -263,7 +263,7 @@ bool R308::serialRead() {
 
   // Verify pack head & sign.校验包头与标识
   if (SerPort.read() == 0xEF && SerPort.read() == 0x01) {
-    for (short i = 0; i < 4; i++) {
+    for (int8_t i = 0; i < 4; i++) {
       if (SerPort.read() != packHead[i + 2])
         return false;
     }
@@ -274,7 +274,7 @@ bool R308::serialRead() {
 
   // Receive pack.接收包
   packSerialRead[0] = ((SerPort.read() << 8) | SerPort.read());
-  for (short i = 1; i <= packSerialRead[0]; i++)
+  for (int8_t i = 1; i <= packSerialRead[0]; i++)
     packSerialRead[i] = SerPort.read();
   return true;
 }
